+++ dropins/class-filter-dropin.php
@@ -26,7 +26,6 @@ class Filter_Dropin extends Dropin {
 		wp_enqueue_style( 'simple_history_FilterDropin', $file_url . 'filter-dropin.css', null, SIMPLE_HISTORY_VERSION );
 	}
 
-
 	public function gui_page_filters() {
 		$loggers_user_can_read = $this->simple_history->get_loggers_that_user_can_read();
 
@@ -540,7 +539,6 @@ class Filter_Dropin extends Dropin {
 
 		$val->gravatar = Helpers::get_avatar( $val->user_email, '18', 'mm' );
 	}
-
 
 	/**
 	 * Print out HTML form date elements for editing post or comment publish date.